# alpenflight/web/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "AlpenFlight API",
    "description" : "Glider club operations platform. Source of truth for the SPA-generated TS client."
  },
  "servers" : [ {
    "url" : "http://localhost:8080"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/v1/aircraft/{id}" : {
      "get" : {
        "tags" : [ "Aircraft" ],
        "summary" : "Read a single Aircraft by id (current state + latest counter inlined).",
        "operationId" : "getAircraft",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "ac-019e30c3-2c00-7001-8000-000000000001",
            "pattern" : "^ac-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Aircraft detail projection.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AircraftDetail"
                }
              }
            }
          },
          "404" : {
            "description" : "No active aircraft with that id.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AircraftDetail"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Aircraft" ],
        "summary" : "Update an Aircraft (owner fields excluded — see transfer-ownership).",
        "operationId" : "updateAircraft",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "ac-019e30c3-2c00-7001-8000-000000000001",
            "pattern" : "^ac-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AircraftUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Updated.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AircraftDetail"
                }
              }
            }
          },
          "404" : {
            "description" : "No active aircraft with that id.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AircraftDetail"
                }
              }
            }
          },
          "409" : {
            "description" : "Immatriculation already in use.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AircraftDetail"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Aircraft" ],
        "summary" : "Soft-delete an Aircraft.",
        "operationId" : "deleteAircraft",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "ac-019e30c3-2c00-7001-8000-000000000001",
            "pattern" : "^ac-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Deleted."
          },
          "404" : {
            "description" : "No active aircraft with that id."
          }
        }
      }
    },
...
    "/api/v1/aircraft" : {
      "get" : {
        "tags" : [ "Aircraft" ],
        "summary" : "List all active aircraft, sorted by immatriculation. Optional `type` query param slices the list per legacy semantics: GLIDER (pure + with-motor), MOTOR (excludes glider-with-motor), TOWING (any type).",
        "operationId" : "listAircraft",
        "parameters" : [ {
          "name" : "type",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "GLIDER", "MOTOR", "TOWING" ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Array of aircraft listitem projections.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/AircraftListItem"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Aircraft" ],
        "summary" : "Register a new Aircraft. Immatriculation must be globally unique.",
        "operationId" : "registerAircraft",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AircraftCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AircraftDetail"
                }
              }
            }
          },
          "400" : {
            "description" : "Validation failed.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AircraftDetail"
                }
              }
            }
          },
          "409" : {
            "description" : "Immatriculation already in use.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AircraftDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/aircraft/{id}/transfer-ownership" : {
      "post" : {
        "tags" : [ "Aircraft" ],
        "summary" : "Transfer ownership of an Aircraft (CLUB_ADMINISTRATOR of the managing club).",
        "operationId" : "transferOwnership",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "ac-019e30c3-2c00-7001-8000-000000000001",
            "pattern" : "^ac-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AircraftTransferOwnershipRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Ownership transferred.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AircraftDetail"
                }
              }
            }
          },
          "404" : {
            "description" : "No active aircraft with that id.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AircraftDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/aircraft/{id}/states" : {
      "get" : {
        "tags" : [ "Aircraft" ],
        "summary" : "List state history for an Aircraft (newest first).",
        "operationId" : "listStateHistory",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "ac-019e30c3-2c00-7001-8000-000000000001",
            "pattern" : "^ac-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "State history.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AircraftStateHistory"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Aircraft" ],
        "summary" : "Record a new airworthiness state (closes the previous open period).",
        "operationId" : "changeState",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "ac-019e30c3-2c00-7001-8000-000000000001",
            "pattern" : "^ac-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AircraftStateChangeRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "State change recorded.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AircraftStateHistoryEntryResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Concurrent modification or invalid validFrom.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AircraftStateHistoryEntryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/aircraft/{id}/counters" : {
      "get" : {
        "tags" : [ "Aircraft" ],
        "summary" : "List counter history for an Aircraft (newest first).",
        "operationId" : "listCounterHistory",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "ac-019e30c3-2c00-7001-8000-000000000001",
            "pattern" : "^ac-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Counter history.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AircraftCounterHistory"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Aircraft" ],
        "summary" : "Record a new operating-counter snapshot.",
        "operationId" : "recordCounter",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "example" : "ac-019e30c3-2c00-7001-8000-000000000001",
            "pattern" : "^ac-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AircraftCounterRecordRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Counter recorded.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AircraftOperatingCounterResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "Non-monotonic totals or duplicate at_date_time.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AircraftOperatingCounterResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "AircraftUpdateRequest" : {
        "type" : "object",
        "description" : "Payload to update an Aircraft. Ownership fields are not settable here — use the transfer-ownership endpoint (A04 defense).",
        "properties" : {
          "aircraftTypeId" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "019e2e15-2c00-7af9-8000-000000002af9",
            "pattern" : "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "immatriculation" : {
            "type" : "string",
            "maxLength" : 15,
            "minLength" : 2,
            "pattern" : "^[A-Z0-9-]{2,15}$"
          },
          "manufacturerName" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          },
          "aircraftModel" : {
            "type" : "string",
            "maxLength" : 50,
            "minLength" : 0
          },
          "competitionSign" : {
            "type" : "string",
            "pattern" : "^[A-Za-z0-9]{1,5}$"
          },
          "flarmId" : {
            "type" : "string",
            "pattern" : "^[A-Fa-f0-9]{6}$"
          },
          "aircraftSerialNumber" : {
            "type" : "string",
            "maxLength" : 20,
            "minLength" : 0
          },
          "yearOfManufacture" : {
            "type" : "string",
            "format" : "date"
          },
          "noiseClass" : {
            "type" : "string",
            "maxLength" : 1,
            "minLength" : 0
          },
          "noiseLevel" : {
            "type" : "number"
          },
          "mtom" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "nrOfSeats" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "flightOperatingCounterUnitTypeId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "engineOperatingCounterUnitTypeId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "homebaseId" : {
            "type" : "string",
            "example" : "loc-019e30c3-2c00-7001-8000-000000000001",
            "pattern" : "^loc-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "spotLink" : {
            "type" : "string",
            "maxLength" : 250,
            "minLength" : 0,
            "pattern" : "^https://.*"
          },
          "isTowingOrWinchRequired" : {
            "type" : "boolean"
          },
          "isTowingStartAllowed" : {
            "type" : "boolean"
          },
          "isWinchStartAllowed" : {
            "type" : "boolean"
          },
          "isTowingAircraft" : {
            "type" : "boolean"
          },
          "comment" : {
            "type" : "string",
            "maxLength" : 250,
            "minLength" : 0
          },
          "daecIndex" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "aircraftTypeId", "immatriculation" ]
      },
      "AircraftDetail" : {
        "type" : "object",
        "description" : "Aircraft detail projection — full payload including current state + latest counter.",
        "properties" : {
          "id" : {
            "type" : "string",
            "example" : "ac-019e30c3-2c00-7001-8000-000000000001",
            "pattern" : "^ac-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "ownerClubId" : {
            "type" : "string",
            "example" : "clb-019e30c3-2c00-7001-8000-000000000001",
            "pattern" : "^clb-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "aircraftTypeId" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "019e2e15-2c00-7af9-8000-000000002af9",
            "pattern" : "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "immatriculation" : {
            "type" : "string"
          },
          "manufacturerName" : {
            "type" : "string"
          },
          "aircraftModel" : {
            "type" : "string"
          },
          "competitionSign" : {
            "type" : "string"
          },
          "flarmId" : {
            "type" : "string"
          },
          "aircraftSerialNumber" : {
            "type" : "string"
          },
          "yearOfManufacture" : {
            "type" : "string",
            "format" : "date"
          },
          "noiseClass" : {
            "type" : "string"
          },
          "noiseLevel" : {
            "type" : "number"
          },
          "mtom" : {
            "type" : "integer",
            "format" : "int32"
          },
          "nrOfSeats" : {
            "type" : "integer",
            "format" : "int32"
          },
          "aircraftOwnerPersonId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "flightOperatingCounterUnitTypeId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "engineOperatingCounterUnitTypeId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "homebaseId" : {
            "type" : "string",
            "example" : "loc-019e30c3-2c00-7001-8000-000000000001",
            "pattern" : "^loc-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "spotLink" : {
            "type" : "string"
          },
          "isTowingOrWinchRequired" : {
            "type" : "boolean"
          },
          "isTowingStartAllowed" : {
            "type" : "boolean"
          },
          "isWinchStartAllowed" : {
            "type" : "boolean"
          },
          "isTowingAircraft" : {
            "type" : "boolean"
          },
          "isFastEntryRecord" : {
            "type" : "boolean"
          },
          "comment" : {
            "type" : "string"
          },
          "daecIndex" : {
            "type" : "integer",
            "format" : "int32"
          },
          "currentState" : {
            "$ref" : "#/components/schemas/AircraftStateHistoryEntryResponse"
          },
          "latestCounter" : {
            "$ref" : "#/components/schemas/AircraftOperatingCounterResponse"
          }
        },
        "required" : [ "aircraftTypeId", "id", "immatriculation", "isFastEntryRecord", "isTowingAircraft", "isTowingOrWinchRequired", "isTowingStartAllowed", "isWinchStartAllowed" ]
      },
...
      "AircraftCreateRequest" : {
        "type" : "object",
        "description" : "Payload to register a new Aircraft. Ownership defaults to the managing club; change via transfer-ownership.",
        "properties" : {
          "aircraftTypeId" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "019e2e15-2c00-7af9-8000-000000002af9",
            "pattern" : "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "immatriculation" : {
            "type" : "string",
            "maxLength" : 15,
            "minLength" : 2,
            "pattern" : "^[A-Z0-9-]{2,15}$"
          },
          "manufacturerName" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 0
          },
          "aircraftModel" : {
            "type" : "string",
            "maxLength" : 50,
            "minLength" : 0
          },
          "competitionSign" : {
            "type" : "string",
            "pattern" : "^[A-Za-z0-9]{1,5}$"
          },
          "flarmId" : {
            "type" : "string",
            "pattern" : "^[A-Fa-f0-9]{6}$"
          },
          "aircraftSerialNumber" : {
            "type" : "string",
            "maxLength" : 20,
            "minLength" : 0
          },
          "yearOfManufacture" : {
            "type" : "string",
            "format" : "date"
          },
          "noiseClass" : {
            "type" : "string",
            "maxLength" : 1,
            "minLength" : 0
          },
          "noiseLevel" : {
            "type" : "number"
          },
          "mtom" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "nrOfSeats" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "flightOperatingCounterUnitTypeId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "engineOperatingCounterUnitTypeId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "homebaseId" : {
            "type" : "string",
            "example" : "loc-019e30c3-2c00-7001-8000-000000000001",
            "pattern" : "^loc-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "spotLink" : {
            "type" : "string",
            "maxLength" : 250,
            "minLength" : 0,
            "pattern" : "^https://.*"
          },
          "isTowingOrWinchRequired" : {
            "type" : "boolean"
          },
          "isTowingStartAllowed" : {
            "type" : "boolean"
          },
          "isWinchStartAllowed" : {
            "type" : "boolean"
          },
          "isTowingAircraft" : {
            "type" : "boolean"
          },
          "comment" : {
            "type" : "string",
            "maxLength" : 250,
            "minLength" : 0
          },
          "daecIndex" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "aircraftTypeId", "immatriculation" ]
      },
      "AircraftTransferOwnershipRequest" : {
        "type" : "object",
        "description" : "CLUB_ADMINISTRATOR-only payload to transfer ownership metadata. Managing tenant is unchanged.",
        "properties" : {
          "newOwnerClubId" : {
            "type" : "string",
            "example" : "clb-019e30c3-2c00-7001-8000-000000000001",
            "pattern" : "^clb-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "newOwnerPersonId" : {
            "type" : "string",
            "format" : "uuid"
          }
        }
      },
...
      "AircraftListItem" : {
        "type" : "object",
        "description" : "Aircraft listitem projection — sorted by immatriculation.",
        "properties" : {
          "id" : {
            "type" : "string",
            "example" : "ac-019e30c3-2c00-7001-8000-000000000001",
            "pattern" : "^ac-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "ownerClubId" : {
            "type" : "string",
            "example" : "clb-019e30c3-2c00-7001-8000-000000000001",
            "pattern" : "^clb-[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "immatriculation" : {
            "type" : "string"
          },
          "competitionSign" : {
            "type" : "string"
          },
          "aircraftTypeId" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "019e2e15-2c00-7af9-8000-000000002af9",
            "pattern" : "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "aircraftTypeCode" : {
            "type" : "string"
          },
          "hasEngine" : {
            "type" : "boolean"
          },
          "isTowingAircraft" : {
            "type" : "boolean"
          },
          "currentStateCode" : {
            "type" : "string"
          },
          "currentStateFlyable" : {
            "type" : "boolean"
          }
        },
        "required" : [ "aircraftTypeCode", "aircraftTypeId", "hasEngine", "id", "immatriculation", "isTowingAircraft" ]
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "OIDC bearer token (Keycloak in dev, hosted IdP TBD in prod).",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}